Add Support for Context Fields
Add new context key to store fields on the context. Fields stored under
this key will be logged with the LogFieldsCtx and LogFieldsCtxD
functions.

Add functions for storing and retrieving fields from the a context, they
will create a new map of fields so the fields stored on the context will
not be changed if the map passed to or returned from these funcs is
changed later.